+++ GLMakie/test/runtests.jl
@@ -109,15 +109,23 @@ end
     sleep(0.1)
     GLMakie.closeall()
 
-        # Why does it start with a skipped tick?
+    # Screen initalizes with `requires_update = false`, so it may skip renders
     i = 1
     while tick_record[i].state == Makie.SkippedRenderTick
         check_tick(tick_record[1], Makie.SkippedRenderTick, i)
         i += 1
     end
 
+    # apply_config!() and scene insertion cause renders
+    # (number of ticks is probably performance sensitive here so require 1, allow 1..3)
     check_tick(tick_record[i], Makie.RegularRenderTick, i)
     i += 1
+    for j in 1:2
+        if tick_record[i].state == Makie.RegularRenderTick
+            check_tick(tick_record[i], Makie.RegularRenderTick, i)
+            i += 1
+        end
+    end
 
     while tick_record[i].state == Makie.SkippedRenderTick
             check_tick(tick_record[i], Makie.SkippedRenderTick, i)
@@ -140,18 +148,26 @@ end
     # texture atlas is triggered by text
     # include SSAO to make sure its cleanup works too
     f,a,p = image(rand(4,4))
+
+    # make sure switching the render pipeline doesn't cause errors
+    screen = display(f, ssao = true, visible = false)
+    colorbuffer(screen)
+    screen = display(f, ssao = false, visible = false)
+    colorbuffer(screen)
     screen = display(f, ssao = true, visible = false)
     colorbuffer(screen)
 
     # verify that SSAO is active
-    @test screen.postprocessors[1] != GLMakie.empty_postprocessor()
+    @test :SSAO1 in map(x -> x.name, screen.render_pipeline.parent.stages)
 
-    framebuffer = screen.framebuffer
-    framebuffer_textures = copy(screen.framebuffer.buffers)
+    framebuffer = screen.framebuffer_factory
+    framebuffer_depth = GLMakie.get_buffer(screen.framebuffer_factory.fb, :depth_stencil)
+    framebuffer_textures = copy(screen.framebuffer_factory.buffers)
+    framebuffer_children = copy(screen.framebuffer_factory.children)
     atlas_textures = first.(values(GLMakie.atlas_texture_cache))
     shaders = vcat([[shader for shader in values(shaders)] for shaders in values(screen.shader_cache.shader_cache)]...)
     programs = [program for program in values(screen.shader_cache.program_cache)]
-    postprocessors = copy(screen.postprocessors)
+    pipeline = copy(screen.render_pipeline.steps)
     robjs = last.(screen.renderlist)
 
     GLMakie.destroy!(screen)
@@ -164,10 +180,25 @@ end
     end
 
     @testset "Framebuffer" begin
-        @test framebuffer.id == 0
+        # GLFramebuffer object
+        @test framebuffer.fb.id == 0
+        @test all(x -> x.id == 0, framebuffer.fb.buffers)
+
+        # FramebufferFactory object
+        @test isempty(framebuffer.children)
+        @test isempty(framebuffer.buffers)
+
         @test !isempty(framebuffer_textures)
-        for (k, tex) in framebuffer_textures
+        for tex in framebuffer_textures
             @test tex.id == 0
+        end
+
+        @test !isempty(framebuffer_children)
+        for fb in framebuffer_children
+            @test fb.id == 0
+            @test framebuffer.fb.id == 0
+            # should automatically be true if all framebuffer_textures == 0
+            @test all(x -> x.id == 0, framebuffer.fb.buffers)
         end
     end
 
@@ -202,10 +233,10 @@ end
     end
 
     @testset "PostProcessors" begin
-        @test map(pp -> length(pp.robjs), postprocessors) == [2,1,2,1]
-        for pp in postprocessors
-            for robj in pp.robjs
-                validate_robj(robj)
+        @test length(pipeline) == 10
+        for step in pipeline
+            if hasfield(typeof(step), :robj)
+                validate_robj(getfield(step, :robj))
             end
         end
     end
@@ -218,5 +249,6 @@ end
     end
 
     # Check that no finalizers triggered on un-freed objects throughout all tests
+    GC.gc(true)
     @test GLMakie.GLAbstraction.FAILED_FREE_COUNTER[] == 0
 end